+++ src/server/lib/signin/oauth.js
@@ -1,4 +1,5 @@
-import { openidClient, oAuth1Client } from "../oauth/client"
+import { openidClient } from "../oauth/client"
+import { oAuth1Client } from "../oauth/oauth1"
 import logger from "../../../lib/logger"
 import { createState } from "../oauth/state-handler"
 import { createPKCE } from "../oauth/pkce-handler"
